+++ GegeIzi/src/gegeizi/Position.java
@@ -12,10 +12,10 @@ import org.json.simple.JSONObject;
  * @author user
  */
 public class Position {
-    private int x;
-    private int y;
+    private long x;
+    private long y;
     
-    public Position(int x, int y){
+    public Position(long x, long y){
         this.y = x;
         this.y = y;
     }
@@ -26,23 +26,23 @@ public class Position {
     
     public Position(JSONObject obj){
         //ints
-        x = (int) obj.get("x");
-        y = (int) obj.get("y");
+        x = (long) obj.get("x");
+        y = (long) obj.get("y");
     }
 
-    public int getX() {
+    public long getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(long x) {
         this.x = x;
     }
 
-    public int getY() {
+    public long getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(long y) {
         this.y = y;
     }
